backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.0.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-decouple==3.8
redis==6.4.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.9.0
zope.interface==7.2
